+++ src/nodeTreeWidget.cpp
@@ -22,6 +22,7 @@
 #include <la/avdecc/controller/internals/avdeccControlledEntity.hpp>
 #include <la/avdecc/logger.hpp>
 #include "avdecc/controllerManager.hpp"
+#include "avdecc/hiveLogItems.hpp"
 #include "avdecc/helper.hpp"
 #include "toolkit/textEntry.hpp"
 #include "toolkit/comboBox.hpp"
@@ -387,7 +388,17 @@ private:
 
 		// Dynamic model
 		{
-			auto* dynamicItem = new AvbInterfaceDynamicTreeWidgetItem(_controlledEntityID, node.descriptorIndex, node.dynamicModel, q);
+			auto linkStatus = la::avdecc::controller::ControlledEntity::InterfaceLinkStatus::Unknown;
+			try
+			{
+				linkStatus = controlledEntity->getAvbInterfaceLinkStatus(node.descriptorIndex);
+			}
+			catch (...)
+			{
+				AVDECC_ASSERT(false, "Should not happen");
+				LOG_HIVE_ERROR(QString("Visit AvbInterfaceNode %1 for %2: Exception while getting AvbInterfaceLinkStatus").arg(node.descriptorIndex).arg(avdecc::helper::uniqueIdentifierToString(controlledEntity->getEntity().getEntityID())));
+			}
+			auto* dynamicItem = new AvbInterfaceDynamicTreeWidgetItem(_controlledEntityID, node.descriptorIndex, node.dynamicModel, linkStatus, q);
 			dynamicItem->setText(0, "Dynamic Info");
 		}